+++ docs/changes.rst
@@ -12,6 +12,7 @@ Released on ? 2014.
 * Added support for JSON file format.
 * Fixed generating mo files in certain cases.
 * Added support for GitLab notifications.
+* Added support for disabling translation suggestions.
 
 weblate 1.9
 -----------